+++ front/src/components/pages/WriteBlogPost/WritePostPresenter.tsx
